+++ lib/unidom/common.rb
@@ -1,8 +1,18 @@
 require 'unidom/common/engine'
 
+require 'unidom/common/numeration'
+
 module Unidom
   module Common
+
     NULL_UUID = '00000000-0000-0000-0000-000000000000'.freeze
     SELF      = '~'.freeze
+
+    #OPENED_AT = Time.utc(1970).freeze
+    #CLOSED_AT = Time.utc(3000).freeze
+
+    #FROM_DATE = '1970-01-01'.freeze
+    #THRU_DATE = '3000-01-01'.freeze
+
   end
 end